[API][DB] Fixed knex setup.
