+++ api/source/graphql/resolvers.ts
@@ -1,6 +1,6 @@
 // (c) Tecnologico de Monterrey 2022, rights reserved.
 
-import db from "../database/database";
+import { db } from "../database/database";
 import { ROLE_TABLE_NAME, USER_TABLE_NAME } from "../database/utils/database_constants";
 
 /**
